requirements.txt
-i https://pypi.org/simple
click==8.0.1; python_version >= '3.6'
colorama==0.4.4; platform_system == 'Windows'
flask-cors==3.0.10
flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
numpy==1.21.2
opencv-python-headless==4.5.3.56
pillow==8.3.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==2.0.1; python_version >= '3.6'
